fix: move carla parity to bridge package
